es-app/GuiMenu: Map brigtness quadratically.
By mapping the brightness to a quadratic function, the lower range of
the hardware values will have a higher range in the user facing slider.
This will make the brightness level appear to change in a linear
fashion. I did not actually measure this on my display, but the apparent
change is now more consistent.

The function is a reduced version of `100 * x^2 / 100^2`.